requirements.txt
streamlit
streamlit-option-menu
sentence-transformers
pinecone-client
google.generativeai
python-dotenv
pymongo
